Use Collectives::SetEnvInfo and rearrange tests
