+++ xla/pjrt/gpu/se_gpu_pjrt_client.cc
@@ -47,6 +47,7 @@ limitations under the License.
 #include "absl/time/time.h"
 #include "absl/types/span.h"
 #include "xla/client/local_client.h"
+#include "xla/core/collectives/collectives_registry.h"
 #include "xla/executable_run_options.h"
 #include "xla/hlo/builder/xla_computation.h"
 #include "xla/layout.h"
@@ -119,7 +120,6 @@ limitations under the License.
 #if GOOGLE_CUDA
 #include "third_party/gpus/cuda/include/cuda.h"
 #include "third_party/gpus/cuda/include/cuda_runtime_api.h"
-#include "xla/backends/gpu/collectives/nvshmem_collectives.h"
 #include "xla/service/gpu/model/gpu_collective_performance_model.h"
 #include "xla/stream_executor/gpu/gpu_cudamallocasync_allocator.h"
 #elif TENSORFLOW_USE_ROCM
@@ -1245,15 +1245,14 @@ absl::StatusOr<DeviceTopologyPair> BuildDistributedDevices(
   }
   gpu_executable_run_options->set_gpu_global_device_ids(
       std::move(gpu_device_ids));
-#if GOOGLE_CUDA
   static const bool xla_gpu_experimental_enable_nvshmem =
       xla::GetDebugOptionsFromFlags().xla_gpu_experimental_enable_nvshmem();
   if (xla_gpu_experimental_enable_nvshmem) {
-    xla::gpu::NvshmemCollectives::Default()->SetEnvInfo(
-        node_id, global_topology.nodes().size(), local_device_states.size(),
-        kv_store);
+    TF_ASSIGN_OR_RETURN(Collectives * collectives,
+                        CollectivesRegistry::Get("gpu", "nvshmem"));
+    collectives->SetEnvInfo(node_id, global_topology.nodes().size(),
+                            local_device_states.size(), kv_store);
   }
-#endif  // GOOGLE_CUDA
 #if GOOGLE_CUDA || TENSORFLOW_USE_ROCM
   if (num_nodes > 1) {
     auto nccl_id_store = std::make_shared<NcclIdStore>(node_id, device_to_node,
