+++ xla/stream_executor/cuda/BUILD
@@ -237,7 +237,7 @@ cc_library(
     ],
     deps = [
         "//xla:debug_options_flags",
-        "//xla/backends/gpu/collectives:nvshmem_collectives",
+        "//xla/core/collectives:collectives_registry",
         "//xla/stream_executor:activate_context",
         "//xla/stream_executor:stream_executor_h",
         "@com_google_absl//absl/log",
